requirements.txt
# This is a requirements file for a Python project.
# It lists the packages required for the project along with their versions.
# To install the packages, run:
# pip install -r requirements.txt
pandas
numpy
scikit-learn
flask
python-dotenv
